+++ StardewVoidEffects/Framework/ModConfig.cs
@@ -1,0 +1,9 @@
+﻿namespace StardewVoidEffects.Framework
+{
+    internal class ModConfig
+    {
+        public bool modEnabledOnStartup { get; set; } = true;
+        public float VoidItemPriceIncrease { get; set; } = 2.0f;
+        public int VoidDecay { get; set; } = 10;
+    }
+}
