+++ StardewVoidEffects/Framework/SavedData.cs
@@ -1,0 +1,7 @@
+﻿namespace StardewVoidEffects.Framework
+{
+    internal class SavedData
+    {
+        public float Tolerance { get; set; }
+    }
+}
